requirements.txt
mysql-connector-python
pandas
requests
streamlit
